+++ src/main/java/com/asd/enumerator/TipoActivoEnum.java
@@ -2,24 +2,24 @@ package com.asd.enumerator;
 
 public enum TipoActivoEnum {
 	
-	TERRENO(1L,"Terrenos y bienes naturales"),
-	CONTRUCCION(2L,"Construcciones"),
-	INSTALACIONES(3L,"Instalaciones técnicas"),
-	MAQUINARIA(4L,"Maquinaria"),
-	MOBILIARIO(5L,"Mobiliario"),
-	EQUIPOS(6L,"Equipos para procesos informáticos"),
-	TRANSPORTE(7L,"Elementos de transporte"),
-	OTROS(8L,"Otros");
+	TERRENO(1,"Terrenos y bienes naturales"),
+	CONTRUCCION(2,"Construcciones"),
+	INSTALACIONES(3,"Instalaciones técnicas"),
+	MAQUINARIA(4,"Maquinaria"),
+	MOBILIARIO(5,"Mobiliario"),
+	EQUIPOS(6,"Equipos para procesos informáticos"),
+	TRANSPORTE(7,"Elementos de transporte"),
+	OTROS(8,"Otros");
 	
-	private Long code;
+	private int code;
 	private String description;
 	
-	TipoActivoEnum(Long code, String description){
+	TipoActivoEnum(int code, String description){
 		this.code = code;
 		this.description = description;
 	}
 
-	public Long getCode() {
+	public int getCode() {
 		return code;
 	}
